Analysis of querydepth through jSqlParser library
Implemented two new classes: `QueryDepthExtractor` and `QueryDepthVisitor`. `QueryDepthExtractor` uses the visitor to analyse the query syntax tree and extract the depth as an integer (but can also be extracted as a list of operations which cause the depth to increment).

There are some assumptions in the classes: the query depth is only incremented by SELECT statements (also subqueries), JOINs and HAVING statements. 

TODO: implement this into EvoSQL by analysing the query before execution and using the querydepth in fitness calculation.